requirements.txt
face_recognition
flask
requests
Pillow
